Use thumbnails

+++ src/build.js
@@ -1,9 +1,11 @@
 const fs = require("fs-extra")
 const lib = require("./lib.js")
 const assembler = require("./assembler.js")
+const child_process = require("child_process");
 
 fs.ensureDirSync(lib.buildDir);
 for (const file of fs.readdirSync(lib.buildDir)) {
+    if (file === "img") continue
     fs.removeSync(lib.buildDir + file);
 }
 //for (const file of fs.readdirSync(lib.staticDir)) { 
@@ -12,8 +14,12 @@ for (const file of fs.readdirSync(lib.buildDir)) {
 
 
 fs.ensureDirSync(lib.imageOutDir);
+fs.ensureDirSync(lib.imageThumbsOutDir);
 for (const file of fs.readdirSync(lib.publicImagesDir)) {
+    if (fs.existsSync(lib.imageOutDir + file)) continue
     fs.link(lib.publicImagesDir + file, lib.imageOutDir + file)
+    child_process.execSync("magick " + lib.publicImagesDir + file + " -gravity center -scale 300x300^ -extent 300x300 " + lib.imageThumbsOutDir + file)
+    console.log("Wrote " + lib.imageThumbsOutDir + file)
 }
 
 assembler.buildCollection("All images", fs.readdirSync(lib.publicImagesDir));
